Feature_VH/backend/requirements.txt

# --- Core Frameworks ---
Django==5.1.5
fastapi==0.115.8

# --- Django Extensions ---
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3

# --- TiDB & Database Clients ---
django-tidb==5.1.0
mysqlclient==2.2.7
PyMySQL==1.1.1
databases[aiomysql]==0.9.0
aiomysql==0.2.0

# --- Render Production Setup ---
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
python-dotenv==1.0.1

# --- Utilities ---
Pillow==11.1.0
uvicorn[standard]==0.34.0
websockets==14.2
python-multipart==0.0.20
PyJWT==2.10.1
cryptography==44.0.0
